+++ src/main/java/com/maxmind/geoip2/record/RepresentedCountry.java
@@ -30,39 +30,19 @@ public final class RepresentedCountry extends Country {
         this(null, null, null, false, null, null, null);
     }
 
+    @MaxMindDbConstructor
     public RepresentedCountry(
-            @JacksonInject("locales") List<String> locales,
-            @JsonProperty("confidence") Integer confidence,
-            @JsonProperty("geoname_id") Integer geoNameId,
-            @JsonProperty("is_in_european_union") boolean isInEuropeanUnion,
-            @JsonProperty("iso_code") String isoCode,
-            @JsonProperty("names") Map<String, String> names,
-            @JsonProperty("type") String type
+            @JacksonInject("locales") @MaxMindDbParameter(name = "locales") List<String> locales,
+            @JsonProperty("confidence") @MaxMindDbParameter(name = "confidence") Integer confidence,
+            @JsonProperty("geoname_id") @MaxMindDbParameter(name = "geoname_id") Long geoNameId,
+            @JsonProperty("is_in_european_union") @MaxMindDbParameter(name = "is_in_european_union") Boolean isInEuropeanUnion,
+            @JsonProperty("iso_code") @MaxMindDbParameter(name = "iso_code") String isoCode,
+            @JsonProperty("names") @MaxMindDbParameter(name = "names") Map<String, String> names,
+            @JsonProperty("type") @MaxMindDbParameter(name = "type") String type
     ) {
         super(locales, confidence, geoNameId, isInEuropeanUnion, isoCode,
                 names);
         this.type = type;
-    }
-
-    @MaxMindDbConstructor
-    public RepresentedCountry(
-            @MaxMindDbParameter(name = "locales") List<String> locales,
-            @MaxMindDbParameter(name = "confidence") Integer confidence,
-            @MaxMindDbParameter(name = "geoname_id") Long geoNameId,
-            @MaxMindDbParameter(name = "is_in_european_union") Boolean isInEuropeanUnion,
-            @MaxMindDbParameter(name = "iso_code") String isoCode,
-            @MaxMindDbParameter(name = "names") Map<String, String> names,
-            @MaxMindDbParameter(name = "type") String type
-    ) {
-        this(
-                locales,
-                confidence,
-                geoNameId != null ? geoNameId.intValue() : null,
-                isInEuropeanUnion != null ? isInEuropeanUnion : false,
-                isoCode,
-                names,
-                type
-        );
     }
 
     public RepresentedCountry(